pyproject.toml
[project]
name = "downtime-aggregator"
version = "0.1.0"
description = ""
authors = [
    {name = "Cristian Ticu",email = "cristian.ticu@eggnita.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "boto3[crt] (>=1.38.3,<2.0.0)",
    "pydantic (>=2.11.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
types-boto3 = {extras = ["essential"], version = "^1.38.3"}
aws-lambda-typing = "^2.20.0"

